Revert "RPi Pico: Enable USB console (CDC ACM)"
This reverts commit e4c4d6a5d98b447f856cf31a77a93ca9fffae040.

Unfortunately, the firmware does not start properly if no USB cable
for the USB Console has been connected, possibly even if there is no
active session in a terminal emulator. It is probably due to "busy"
logging buffers and thus a high load in the RPi Pico specific high
priority task for the USB device. It is precisely at this point that
corrections were first introduced with Zephyr v4.1, probably to
prevent this type of early blockage.

Signed-off-by: Stephan Linz <linz@li-pro.net>